+++ CHANGES.md
@@ -10,6 +10,12 @@ NEXT MILESTONE
 ### Major features:
 - Add `Solver.tableSampling` resolution policy which returns a stream of `Solution` randomly sampled.
 - Extend some black-box strategies to `SetVar` and `GraphVar`
+- Add `union(SetVar unionSet, SetVar indices, SetVar[] sets)` constraint
+- Add `increasing` and `decreasing` constraints
+- revamp `pow` constraint
+- increase interval for MiniZinc unbounded intvar
+- Add threshold for Element with array of variables signature
+- 
  
 ### Deprecated API (to be removed in next release):